requirements.txt
Flask
flask-cors
nltk
requests
inflect
# From here all dependencies for ml
scipy
sklearn
pytorch
numpy
pandas
torchvision
xgboost


